% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{ds.hpc.capabilities}
\alias{ds.hpc.capabilities}
\title{Report dsHPC server capabilities}
\usage{
ds.hpc.capabilities(conns)
}
\arguments{
\item{conns}{DSI connections object.}
}
\value{
A \code{dshpc_result} with per-site capability lists.
}
\description{
Report dsHPC server capabilities
}
