% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{ds.hpc.logs}
\alias{ds.hpc.logs}
\title{Tail sanitized dsHPC job logs}
\usage{
ds.hpc.logs(conns, job_id, last_n = 50L)
}
\arguments{
\item{conns}{DSI connections object.}

\item{job_id}{Character job id.}

\item{last_n}{Integer number of log lines requested from each site.}
}
\value{
A \code{dshpc_result} with per-site character vectors.
}
\description{
Tail sanitized dsHPC job logs
}
