% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{ds.hpc.outputs}
\alias{ds.hpc.outputs}
\title{List outputs registered for a dsHPC job}
\usage{
ds.hpc.outputs(conns, job_id)
}
\arguments{
\item{conns}{DSI connections object.}

\item{job_id}{Character job id.}
}
\value{
A \code{dshpc_result} with one output metadata data.frame per site.
}
\description{
List outputs registered for a dsHPC job
}
