% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{ds.hpc.result}
\alias{ds.hpc.result}
\title{Fetch the disclosure-safe dsHPC job result}
\usage{
ds.hpc.result(conns, job_id)
}
\arguments{
\item{conns}{DSI connections object.}

\item{job_id}{Character job id.}
}
\value{
A \code{dshpc_result} with one result object per site.
}
\description{
Fetch the disclosure-safe dsHPC job result
}
