% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{ds.hpc.status}
\alias{ds.hpc.status}
\title{Get dsHPC job status}
\usage{
ds.hpc.status(conns, job_id)
}
\arguments{
\item{conns}{DSI connections object.}

\item{job_id}{Character job id.}
}
\value{
A \code{dshpc_result} with one status list per site.
}
\description{
Get dsHPC job status
}
