% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit.R
\name{ds.hpc.submit}
\alias{ds.hpc.submit}
\title{Submit a job to all nodes}
\usage{
ds.hpc.submit(conns, job)
}
\arguments{
\item{conns}{DSI connections object.}

\item{job}{A dshpc_job object.}
}
\value{
A dshpc_submission with job_id and per-server details.
}
\description{
Submit a job to all nodes
}
