% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{ds.hpc.summary}
\alias{ds.hpc.summary}
\title{Print a formatted summary of all jobs}
\usage{
ds.hpc.summary(conns, label = NULL)
}
\arguments{
\item{conns}{DSI connections object.}

\item{label}{Character or NULL; filter by label.}
}
\description{
Print a formatted summary of all jobs
}
